+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Configure which tabs are present and their order.
+
 ### Changed
 
 - No longer exits on PipeWire errors in release builds.